+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - CodeCaller value builder (moved from ganbarodigital/php-reflection)
 - ExceptionCaller trait
 - ExceptionMessageData trait (in the Traits folder)
+- UnsupportedType trait
 
 ## 1.0.0 - Sat Jun 27 2015